Convert history.js to history.tsx (#32)
* Converted history presentational component to TypeScript

* Changed filter constants to TypeScript

* Let TS infer types when it can

* Separated History Data to an interface and let TS infer types for variables in filterButton

* Got rid of unnecessary ?:

* Renamed historyConst.tsx to historyConst.ts

* Removed render return type in FilterBUtton

* Moved interfaces to its own file

* Merge master
