+++ package.json
@@ -43,9 +43,9 @@
     "@types/node": "^10.11.2",
     "@types/react": "^16.4.14",
     "@types/react-dom": "^16.0.8",
+    "@types/recharts": "^1.0.29",
     "awesome-typescript-loader": "^5.2.1",
     "css-loader": "^1.0.0",
-    "eslint": "^5.5.0",
     "html-webpack-plugin": "^3.2.0",
     "husky": "^0.14.3",
     "interpolate-html-plugin": "^3.0.0",
